--- packages/wallet-backend-update/src/account/schemas/account.schema.ts
@@ -1,6 +1,0 @@
-import { z } from 'zod'
-
-export const accountSchema = z.object({
-  name: z.string(),
-  assetRafikiId: z.string().uuid()
-})
