--- packages/wallet-backend-update/src/rapyd/wallet/schemas/profile.schema.ts
@@ -1,6 +1,0 @@
-import { z } from 'zod'
-
-export const profileSchema = z.object({
-  firstName: z.string({ required_error: 'First name is required' }),
-  lastName: z.string({ required_error: 'Last name is required' })
-})
